My Portfolio Website
I wanted to make a CV site for myself and I created a simple CV site for myself using C#, .NET and MVC. The site consists of two components. One of these is the front side of the site, which can be accessed through the link of the site. The other component is the admin panel. On this site, there are areas such as my skills, interests, and certificates. You can visit my website to get to know me better. https://mberkiskal.com.tr/